Clean up What's On controller, split details from lists
