+++ frontend/app/controllers/WhatsOn.scala
@@ -2,7 +2,9 @@ package controllers
 
 import com.github.nscala_time.time.Imports._
 import configuration.CopyConfig
-import model.RichEvent.CalendarMonthDayGroup
+import model.RichEvent
+import model.RichEvent.MasterclassEvent._
+import model.RichEvent._
 import model._
 import play.api.mvc.Controller
 import services._
@@ -14,28 +16,82 @@ trait WhatsOn extends Controller with ActivityTracking {
   val localEvents: EventbriteService
   val masterclassEvents: EventbriteService
 
-  private def collectAllEvents = {
-    guLiveEvents.events ++ localEvents.events ++ masterclassEvents.events
+  private def allEvents = {
+    guLiveEvents.events ++ localEvents.events
   }
 
-  def calendarGrid = GoogleAuthenticatedStaffAction { implicit request =>
+  private def allEventsByLocation(location: String) = {
+    guLiveEvents.getEventsByLocation(location) ++ localEvents.getEventsByLocation(location)
+  }
+
+  private def allEventsInArchive = {
+    guLiveEvents.getEventsArchive.toList.flatten ++ localEvents.getEventsArchive.toList.flatten
+  }
+
+
+  def list = CachedAction { implicit request =>
     val pageInfo = PageInfo(
       CopyConfig.copyTitleEvents,
       request.path,
       Some(CopyConfig.copyDescriptionEvents)
     )
-    val eventsGroupedByMonth = RichEvent.groupEventsByMonth(collectAllEvents)
-    Ok(views.html.whatson.calendarGrid(eventsGroupedByMonth, pageInfo))
+    val locationOpt = request.getQueryString("location").filter(_.trim.nonEmpty)
+
+    val featuredEvents = EventGroup("Featured", guLiveEvents.getFeaturedEvents)
+    val partnerEvents =  EventGroup("Programming partner events", guLiveEvents.getPartnerEvents)
+    val events = EventGroup("What's on", chronologicalSort(locationOpt.fold(allEvents)(allEventsByLocation)))
+
+    val locationFitlers = getCitiesWithCount(allEvents).map(FilterItem.tupled)
+
+    Ok(views.html.event.eventsList(
+      pageInfo,
+      events,
+      featuredEvents,
+      partnerEvents,
+      locationFitlers,
+      locationOpt
+    ))
   }
 
-  def calendarList = GoogleAuthenticatedStaffAction { implicit request =>
+  def calendar = GoogleAuthenticatedStaffAction { implicit request =>
+    val calendarEvents =
+      CalendarMonthDayGroup("Calendar", groupEventsByDayAndMonth(allEvents))
+
+    Ok(views.html.event.calendar(
+      PageInfo(s"${calendarEvents.title} | Events", request.path, None),
+      calendarEvents
+    ))
+  }
+
+  def listArchive = CachedAction { implicit request =>
+    val calendarArchive =
+      CalendarMonthDayGroup("Archive", groupEventsByDayAndMonth(allEventsInArchive)(implicitly[Ordering[LocalDate]].reverse))
+
+    Ok(views.html.event.eventsListArchive(
+      PageInfo(s"${calendarArchive.title} | Events", request.path, None),
+      calendarArchive
+    ))
+  }
+
+  def masterclassesList = CachedAction { implicit request =>
     val pageInfo = PageInfo(
-      CopyConfig.copyTitleEvents,
+      CopyConfig.copyTitleMasterclasses,
       request.path,
-      Some(CopyConfig.copyDescriptionEvents)
+      Some(CopyConfig.copyDescriptionMasterclasses)
     )
-    val events = CalendarMonthDayGroup("Calendar", RichEvent.groupEventsByDayAndMonth(collectAllEvents))
-    Ok(views.html.whatson.calendarList(pageInfo, events))
+    val eventGroup = EventGroup("Masterclasses", masterclassEvents.events)
+    Ok(views.html.event.masterclassesList(pageInfo, eventGroup))
+  }
+
+  def masterclassesListFilteredBy(rawTag: String, rawSubTag: String = "") = CachedAction { implicit request =>
+    val pageInfo = PageInfo(
+      CopyConfig.copyTitleMasterclasses,
+      request.path,
+      Some(CopyConfig.copyDescriptionMasterclasses)
+    )
+    val tag = decodeTag( if(rawSubTag.nonEmpty) rawSubTag else rawTag )
+    val eventGroup = EventGroup("Masterclasses", masterclassEvents.getTaggedEvents(tag))
+    Ok(views.html.event.masterclassesList(pageInfo, eventGroup, decodeTag(rawTag), decodeTag(rawSubTag)))
   }
 
 }
