+++ frontend/app/views/event/thankyou.scala.html
@@ -41,7 +41,7 @@
                 <h2 class="page-section__headline">Find more events</h2>
             </div>
             <div class="page-section__content">
-                <a class="action u-no-top-margin" href="@routes.Event.list">See all events</a>
+                <a class="action u-no-top-margin" href="@routes.WhatsOn.list">See all events</a>
             </div>
         </section>
