+++ frontend/app/views/index.scala.html
@@ -149,7 +149,7 @@
                                 """Entertaining discussions, debates and interviews with
                                     | Guardian journalists and prominent cultural figures""".stripMargin,
                                 eventCollection.live,
-                                Some(BasicLink(routes.Event.list.url, "All upcoming events"))
+                                Some(BasicLink(routes.WhatsOn.list.url, "All upcoming events"))
                             )
                         </li>
                         <li class="grid__item">
@@ -157,7 +157,7 @@
                                 "Guardian Local",
                                 "Smaller events celebrating local art, food and culture featured in the Guardian",
                                 eventCollection.local,
-                                Some(BasicLink(routes.Event.list.url, "All upcoming events"))
+                                Some(BasicLink(routes.WhatsOn.list.url, "All upcoming events"))
                             )
                         </li>
                         <li class="grid__item">
@@ -165,7 +165,7 @@
                                 "Guardian Masterclasses",
                                 "A range of courses and workshops taught by award-winning professionals",
                                 eventCollection.masterclasses,
-                                Some(BasicLink(routes.Event.masterclassesList.url, "All upcoming masterclasses"))
+                                Some(BasicLink(routes.WhatsOn.masterclassesList.url, "All upcoming masterclasses"))
                             )
                         </li>
                     </ul>
