+++ frontend/app/views/joiner/thankyou.scala.html
@@ -119,7 +119,7 @@
                     }
                     case _ => {
                         <p>This season of events is packed with discussions, debates and interviews you’ll love.  We’ll send you a welcome email shortly to help you get the most from your membership. Why not take a few moments to find out the latest from membership</p>
-                        <a class="action" href="@routes.Event.list">Get started</a>
+                        <a class="action" href="@routes.WhatsOn.list">Get started</a>
                     }
                 }
             }
