+++ frontend/app/views/welcome.scala.html
@@ -37,14 +37,14 @@
                         @featurePanel(
                             Asset.at("images/promo/promo-events.jpg").toString,
                             "Browse our events programme and book tickets now",
-                            routes.Event.list.toString
+                            routes.WhatsOn.list.toString
                         )
                     </li>
                     <li class="grid__item">
                         @featurePanel(
                             Asset.at("images/promo/promo-masterclasses.jpg").toString,
                             "Learn something new with Guardian Masterclasses",
-                            routes.Event.masterclassesList.toString
+                            routes.WhatsOn.masterclassesList.toString
                         )
                     </li>
                     <li class="grid__item">
